fix(endnote): separator noteLine length (i16->i32) + note-area position (Hancom parity)
length: HWPX noteLine="14692344"(한컴 미주 전폭 기본값)이 separator_length(i16)에
0x2ff8=12280 으로 절단돼 짧은 구분선(측정 188px/20.7%)이 나왔다. 필드/파서/PageItem/
직렬화를 i32 로 넓혀 전체 값을 보존 → 렌더러가 content 폭으로 clamp = 전폭(측정
650px/71.4%, 한컴 일치). HWP5 는 i16 슬롯이라 하위 16비트로 기록(포맷 한계).

position: 미주 영역 시작을 직전 본문 문단 마지막 줄 trailing 줄간격(약 600HWPUNIT)만큼
위로 올린다 — 한컴은 그 trailing 줄간격을 note 영역에 포함하지 않는다(측정 y=213 -> 204,
한컴 y=204 일치).

+++ src/renderer/pagination.rs
@@ -326,8 +326,8 @@ pub enum PageItem {
     },
     /// 미주 영역 시작 구분선
     EndnoteSeparator {
-        /// 구분선 길이 (HWP 단위)
-        separator_length: i16,
+        /// 구분선 길이 (HWP 단위). 한컴 전폭 sentinel(14692344)이 i16을 넘으므로 i32.
+        separator_length: i32,
         /// 구분선 위 여백 (HWP 단위)
         margin_above: i16,
         /// 구분선 아래 여백 (HWP 단위)
